Mathematically correct?
For now it seems spacing should NOT be incorporated into the Poisson solver of the grid. However, it NEEDS to be taken into account when calculating heightmap. Offsetting div doesn't seem to affect result, but since it is not justified, the offset part is deactivated. The final height difference is ~1.9mm for 500*500
Spacing system modified, spacing_x and spacing_y will be forced to be the same in the future, meaning that only one of width and height can be specified.